Decorator

+++ PlayGame.cpp
@@ -137,7 +137,7 @@ PlayGame::PlayGame(const std::shared_ptr<Player>& player) {
     this->guessBox.setOutlineThickness(1);
 
 
-    this->inputWord = new PlayerInput<unsigned short >(22, sf::Color::Red, false);
+    this->inputWord = new PlayerInput<double>(22, sf::Color::Red, false);
     this->inputWord->setFont(font);
     this->inputWord->setPosition(sf::Vector2f(170, 310));
 
@@ -444,7 +444,7 @@ PlayGame::PlayGame() {
     this->guessBox.setOutlineThickness(1);
 
 
-    this->inputWord = new PlayerInput<unsigned short>(22, sf::Color::Red, false);
+    this->inputWord = new PlayerInput<double>(22, sf::Color::Red, false);
     this->inputWord->setFont(font);
     this->inputWord->setPosition(sf::Vector2f(170, 310));
